+++ src/source.files/AgAppBinder.ts
@@ -1,0 +1,13 @@
+interface UpdateScratchPad {
+  jsonData: string;
+}
+interface LatestScratchPad {
+  jsonData: string;
+}
+interface FileBrowserData {
+  jsonData: string;
+}
+interface UpdateFileBrowserData {
+  jsonData: string;
+}
+interface UploadBinderResults {}