+++ src/source.files/Squawks.ts
@@ -1,0 +1,7 @@
+import { AgAirplaneSquawks } from "@prisma/client";
+import { Serialize } from "../type-helpers/serializer";
+
+type GetSquawksResult = {
+  IsOpenSquawks: boolean;
+  squawks: Serialize<AgAirplaneSquawks>[];
+};